requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
openai>=1.3.0
python-dotenv>=1.0.0
pypdf2>=3.0.1
faiss-cpu>=1.10.0
tiktoken>=0.9.0
